third_party/blink/web_tests/platform/generic/fast/forms/text/text-reset-click-delete-text-change-event-expected.txt

In a text field after reset when whole text is deleted i.e. text value is empty. When out of focus, it should dispatch change event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Press reset button
PASS text.value is "value"
PASS changeEventCounter is 0
Delete the whole text
Out of focus from text field
PASS text.value is ""
PASS changeEventCounter is 1
PASS successfullyParsed is true

TEST COMPLETE

 
